feat(examples): fold lifecycle examples into the guide (phase 3)
Fourth phase-3 cluster. The 4 standalone example pages for lifecycle hooks,
disposal, cancellation, and the background-service feed (/lifecycle, /disposal,
/cancellation, /background) are removed; each is now an inline live demo in the
Lifecycle guide (docs/lifecycle.md), which gains Disposal and Background service
sections.

- Register the 7 lifecycle demos in DemoRegistry (embedding the probe source —
  the teaching artifact — while Result mounts the live widget).
- Add markers throughout docs/lifecycle.md, plus new "Disposal", "OnUnmount vs
  IDisposable", and "Background service" sections; drop the stale page-file
  pointer at the end.
- Promote the mount/unmount widgets that lived inline in the pages into their own
  demo components (LifecycleCycleDemo, DisposalSync/Async/UnmountDemo,
  CancellationDemo, BackgroundMetricsDemo); the lifecycle probes are unchanged.
  A shared DisposalDemoLog helper renders the three disposal logs.
- Delete the 4 *Page.cs; drop their ShowcaseLayout links and HomePage cards.
- Retarget LifecyclePageTests -> LifecycleCycleDemoTests; drop the 4 lifecycle
  rows from PageBaselineTests and the removed cards from HomePageTests.
- Consolidate the E2E into a new WalkLifecycleGuideAsync that opens the guide
  once, hydration-gates on the last demo, and drives every demo by #id.
