+++ samples/Rask.Example.Shared/Features/Home/HomePage.cs
@@ -22,12 +22,9 @@ public sealed class HomePage(Navigator nav) : Component
         ("Components", "bi-signpost-2", "Routing", "Attributes, nested layouts, Outlet().", "/routing"),
         ("Components", "bi-link-45deg", "Route + query params", "Bind URL segments and query.", "/users/42"),
         ("Components", "bi-compass", "Navigator", "Programmatic navigation.", "/navigator"),
-        ("Components", "bi-arrow-repeat", "Lifecycle", "Mount, props-changed, rendered hooks.", "/lifecycle"),
         ("Components", "bi-bell", "Toast", "Show, stack, dismiss & auto-hide — no JS.", Routes.ToastPage()),
         ("Components", "bi-megaphone", "Flash messages", "Rails-style transient messages via IFlash.", Routes.FlashPage()),
         ("Components", "bi-bullseye", "Element refs", "Reach the live DOM from C#.", "/element-ref"),
-        ("Components", "bi-x-circle", "Cancellation", "A token that fires on unmount.", "/cancellation"),
-        ("Components", "bi-trash", "Disposal", "IDisposable / IAsyncDisposable cleanup.", "/disposal"),
         ("Components", "bi-mouse", "Events", "DOM event handlers.", "/events"),
         ("Components", "bi-table", "Data table", "Sortable, paginated table.", "/table"),
         ("Components", "bi-list-nested", "Master-detail", "Collapsible rows with a nested datagrid.", "/master-detail"),
