+++ samples/Rask.Example.Shared/Shared/ShowcaseLayout.cs
@@ -32,15 +32,11 @@ public sealed class ShowcaseLayout(RouteState route, IEnumerable<ShowcaseNavEntr
         (Features.Routes.RoutingPage(), "Routing", "bi-signpost-2", "Components", null),
         (Features.Routes.UserDetailPage("42"), "Route + query params", "bi-link-45deg", "Components", "/users"),
         (Features.Routes.NavigatorPage(), "Navigator", "bi-compass", "Components", null),
-        (Features.Routes.LifecyclePage(), "Lifecycle", "bi-arrow-repeat", "Components", null),
         (Features.Routes.ToastPage(), "Toast", "bi-bell", "Components", null),
         (Features.Routes.FlashPage(), "Flash messages", "bi-megaphone", "Components", null),
         (Features.Routes.ElementRefPage(), "Element refs", "bi-bullseye", "Components", null),
         (Features.Routes.UserPage(), "User & auth", "bi-person-lock", "Components", null),
         (Features.Routes.LiveTickerPage("BTC"), "Live ticker", "bi-graph-up-arrow", "Components", "/realtime"),
-        (Features.Routes.BackgroundServicePage(), "Background service", "bi-broadcast", "Components", null),
-        (Features.Routes.CancellationPage(), "Cancellation", "bi-x-circle", "Components", null),
-        (Features.Routes.DisposalPage(), "Disposal", "bi-trash", "Components", null),
         (Features.Routes.EventsPage(), "Events", "bi-mouse", "Components", null),
         (Features.Routes.TablePage(), "Data table", "bi-table", "Components", null),
         (Features.Routes.OrdersPage(), "Master-detail", "bi-list-nested", "Components", null),
