+++ tests/Rask.Example.Shared.Tests/Pages/PageBaselineTests.cs
@@ -11,14 +11,10 @@ public sealed class PageBaselineTests
 {
     [Theory]
     [InlineData(typeof(HomePage), "/", "Welcome")]
-    [InlineData(typeof(BackgroundServicePage), "/background", "Background service")]
-    [InlineData(typeof(CancellationPage), "/cancellation", "Cancellation")]
     [InlineData(typeof(ComponentsPage), "/components", "User components")]
-    [InlineData(typeof(DisposalPage), "/disposal", "Disposal")]
     [InlineData(typeof(DownloadPage), "/download", "File download")]
     [InlineData(typeof(EventsPage), "/events", "Events")]
     [InlineData(typeof(HttpPage), "/http", "HttpClient")]
-    [InlineData(typeof(LifecyclePage), "/lifecycle", "Lifecycle")]
     [InlineData(typeof(LiveTickerPage), "/realtime/BTC", "BTC live ticker")]
     [InlineData(typeof(NavigatorPage), "/navigator", "Navigator")]
     [InlineData(typeof(PrimitivesPage), "/primitives", "Primitives")]
@@ -46,14 +42,10 @@ public sealed class PageBaselineTests
 
     [Theory]
     [InlineData(typeof(HomePage))]
-    [InlineData(typeof(BackgroundServicePage))]
-    [InlineData(typeof(CancellationPage))]
     [InlineData(typeof(ComponentsPage))]
-    [InlineData(typeof(DisposalPage))]
     [InlineData(typeof(DownloadPage))]
     [InlineData(typeof(EventsPage))]
     [InlineData(typeof(HttpPage))]
-    [InlineData(typeof(LifecyclePage))]
     [InlineData(typeof(LiveTickerPage))]
     [InlineData(typeof(NavigatorPage))]
     [InlineData(typeof(NotFoundPage))]
